% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_diffs}
\alias{get_diffs}
\title{Extract the value-differences data frame}
\usage{
get_diffs(result)
}
\arguments{
\item{result}{An object of class \code{"rDataRecon_result"}.}
}
\value{
A \code{data.frame} with columns \code{obs_id}, \code{variable}, \code{base_value},
\code{compare_value}, and \code{diff}.
}
\description{
Pulls \verb{$value_diffs} from a \code{\link[=recon]{recon()}} result, with an informative error
when the data were not retained.
}
\examples{
base_df <- data.frame(id = 1:3, score = c(10, 20, 30))
comp_df <- data.frame(id = 1:3, score = c(10, 99, 30))
r <- recon(base_df, comp_df, id = "id")
get_diffs(r)

}
