% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.rDataRecon_result}
\alias{print.rDataRecon_result}
\title{Print method for rDataRecon_result objects}
\usage{
\method{print}{rDataRecon_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"rDataRecon_result"}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Prints a concise summary of a \code{\link[=recon]{recon()}} result.
}
\examples{
base_df <- data.frame(id = 1:3, x = c(1, 2, 3))
comp_df <- data.frame(id = 1:3, x = c(1, 9, 3))
r <- recon(base_df, comp_df, id = "id")
print(r)

}
